% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_ll.R
\name{filter_ll}
\alias{filter_ll}
\title{filter_ll}
\usage{
filter_ll(y, mu0, C0, V, W)
}
\arguments{
\item{y}{- observed time seris}

\item{mu0}{- starting point for state}

\item{C0}{- starting variance for state}

\item{V}{- observation variance}

\item{W}{- evolution variance}
}
\value{
a - evolution state mean

R - evolution state variance

f - predictive dist mean

Q - predictive dist variance

mu - filtered state mean

C - filtered state variance
}
\description{
This function simulates data from a DLM with a local level
}
