% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_locallevel.R
\name{sim_locallevel}
\alias{sim_locallevel}
\title{sim_locallevel}
\usage{
sim_locallevel(num.pts, V, W, mu.0)
}
\arguments{
\item{num.pts}{- number of time points}

\item{V}{- observation variance}

\item{W}{- evolution variance}

\item{mu.0}{- starting point for state: default is 0}
}
\value{
y - observed value

mu - latent state
}
\description{
This function simulates data from a DLM with a local level
}
